requirements.txt
streamlit
opencv-python-headless
numpy
pillow
tensorflow-cpu==2.12.0
